refactor: app and modules
