--- src/app.module.ts
@@ -1,15 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { HealthCheckModule } from './health-check/health-check.module';
-
-@Module({
-  imports: [
-    TypeOrmModule.forRoot(),
-    HealthCheckModule,
-  ],
-  controllers: [AppController],
-  providers: [AppService],
-})
-export class AppModule {}
