+++ src/app/modules/health-check/models/health-check-spec.entity.ts
